refactor(controller): move test helper func to runtime_test.go
