+++ internal/controller/runtime_test.go
@@ -1,0 +1,26 @@
+package controller
+
+// containsArg reports whether args contains the given flag. When value is
+// non-empty, it also requires the immediately following entry to equal value
+// (i.e. `--flag value` as separate slice elements, which is how BuildArgs
+// emits everything).
+func containsArg(args []string, flag, value string) bool {
+	for i, a := range args {
+		if a != flag {
+			continue
+		}
+		if value == "" {
+			return true
+		}
+		if i+1 < len(args) && args[i+1] == value {
+			return true
+		}
+	}
+	return false
+}
+
+// ptrString, ptrBool, ptrInt32 are local helpers so tests read naturally.
+func ptrBool(b bool) *bool          { return &b }
+func ptrFloat64(f float64) *float64 { return &f }
+func ptrInt32(i int32) *int32       { return &i }
+func ptrString(s string) *string    { return &s }
