+++ internal/controller/runtime_vllm_test.go
@@ -25,31 +25,6 @@ import (
 	inferencev1alpha1 "github.com/defilantech/llmkube/api/v1alpha1"
 )
 
-// containsArg reports whether args contains the given flag. When value is
-// non-empty, it also requires the immediately following entry to equal value
-// (i.e. `--flag value` as separate slice elements, which is how BuildArgs
-// emits everything).
-func containsArg(args []string, flag, value string) bool {
-	for i, a := range args {
-		if a != flag {
-			continue
-		}
-		if value == "" {
-			return true
-		}
-		if i+1 < len(args) && args[i+1] == value {
-			return true
-		}
-	}
-	return false
-}
-
-// ptrString, ptrBool, ptrInt32 are local helpers so tests read naturally.
-func ptrString(s string) *string    { return &s }
-func ptrBool(b bool) *bool          { return &b }
-func ptrInt32(i int32) *int32       { return &i }
-func ptrFloat64(f float64) *float64 { return &f }
-
 // TestVLLMBuildArgs is the single table-driven test that covers every new
 // agentic-coding flag and the "not emitted when unset/false" counterpart.
 // Each row asserts a set of must-contain and must-not-contain flags on the
